requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pycparser==2.20
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
